requirements.txt
Django==4.2.7
psycopg[binary]==3.1.12
redis==5.0.1
djangorestframework==3.14.0
celery==5.3.6
django-celery-beat==2.5.0
djangorestframework-simplejwt==5.3.1
